docs(sdds-acore): Fixes after review
