+++ tokens/plasma.sd.service.compose/docs/src/test/kotlin/com/sdds/plasma/sd/service/ComposeDocScreenshotTest.kt
@@ -1,12 +1,8 @@
 package com.sdds.plasma.sd.service
 
-import androidx.compose.runtime.SideEffect
-import androidx.compose.ui.test.onRoot
 import com.github.takahirom.roborazzi.RobolectricDeviceQualifiers
-import com.github.takahirom.roborazzi.captureRoboImage
 import com.sdds.compose.uikit.fixtures.SDK_NUMBER
 import com.sdds.compose.uikit.fixtures.samples.RoborazziConfigDocs
-import org.junit.AfterClass
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.ParameterizedRobolectricTestRunner
@@ -20,48 +16,17 @@ class ComposeDocScreenshotTest(composableSample: ComposableSampleInfo) : Roboraz
 
     @Test
     fun docs_screenshot_test() {
-        var shouldCapture = true
-
         try {
             composeTestRule.content {
-                val providedStyles = LocalProvidedStyles.current
-                SideEffect {
-                    val componentName = composableSample.id
-                        .substringAfterLast(".")
-                        .substringBefore("_")
-                        .lowercase()
-                    shouldCapture = providedStyles.any { styleKey ->
-                        componentName == styleKey.lowercase()
-                    }
-                    println(" componentName: $componentName, ищем в: $providedStyles, совпали: $shouldCapture")
-                }
                 composableSample.sample.reference.invoke()
             }
             composeTestRule.waitForIdle()
-            if (shouldCapture) {
-                composeTestRule.onRoot().captureRoboImage()
-            } else {
-                skippedTests.add(composableSample.id)
-                println("Skipped tests: ${composableSample.id}")
-            }
+            ifNeedScreenshot()
         } catch (e: Throwable) {
             throw AssertionError(
                 "Screenshot test failed: \n${composableSample.id}",
                 e,
             )
-        }
-    }
-
-    companion object {
-        private val skippedTests = mutableListOf<String>()
-
-        @AfterClass
-        @JvmStatic
-        fun skippedTests() {
-            if (skippedTests.isNotEmpty()) {
-                println("Skipped tests: ${skippedTests.size} screenshots:")
-                skippedTests.forEach { println(" - $it") }
-            }
         }
     }
 }
