+++ tokens/plasma.sd.service.view/docs/override-docs/static/screenshots-docusaurus/com_sdds_uikit_fixtures_samples_notificationcontent_NotificationContent_Simple.png
